Revert "Refactor: PlaceSearch 코드 리팩토링"
This reverts commit 4edce118b5512d8e2af88c8c63daedfd341c84fb.
