+++ GIL/GIL/Screen/PlaceSearch/Delegate/PlaceSearchViewController+UICollectionView.swift
@@ -1,0 +1,18 @@
+//
+//  PlaceSearchViewController+UICollectionView.swift
+//  GIL
+//
+//  Created by 송우진 on 4/30/24.
+//
+
+import UIKit
+
+extension PlaceSearchViewController: UICollectionViewDelegate {
+    func collectionView(
+        _ collectionView: UICollectionView,
+        didSelectItemAt indexPath: IndexPath
+    ) {
+        let place = viewModel.mapItems[indexPath.row]
+        viewModel.storePlace(place)
+    }
+}
