+++ GIL/GIL/Screen/PlaceSearch/Extension/PlaceSearchViewController+Actions.swift
@@ -1,0 +1,14 @@
+//
+//  PlaceSearchViewController+Actions.swift
+//  GIL
+//
+//  Created by 송우진 on 5/1/24.
+//
+
+import Foundation
+
+extension PlaceSearchViewController {
+    func backButtonTapped() {
+        navigationController?.popViewController(animated: true)
+    }
+}
