+++ GIL/GIL/Screen/PlaceSearch/Views/PlaceSearchNavigationBar.swift
@@ -57,6 +57,7 @@ extension PlaceSearchNavigationBar {
         searchBar.layer.borderColor = UIColor.lightGray.cgColor
         searchBar.layer.borderWidth = 1
         searchBar.searchTextField.borderStyle = .none
+        
         searchBar
             .top(equalTo: backButton.bottomAnchor, constant: 10)
             .left(equalTo: leadingAnchor, constant: 16)
